#4 - Adding placeholders for spanish and catalan
