+++ src/handlers/languages/catalan.js
@@ -1,0 +1,30 @@
+const labels = {
+  PATIENT_ID: 'Identificación del paciente',
+  DOB: 'Fecha de nacimiento',
+  AGE: 'Edad',
+  GENDER: 'Sexo',
+  BMI: 'IMC',
+  RECORDING_DETAILS: 'Datos de la grabación',
+  DEVICE: 'Dispositivo',
+  RECORDING: 'Grabando',
+  MONITORING_TIME_FLOW: 'Evaluación del flujo',
+  FLOW_EVALUATION: 'Flow evaluation',
+  OXYGEN_SATURATION_EVALUATION: 'Evaluación saturación',
+  EVENTS_INDEX: 'Índice de eventos',
+  SUPINE: 'Supino',
+  NON_SUPINE: 'No supino',
+  UPRIGHT: 'Erguido',
+  EVENTS_TOTAL: 'Eventos totales',
+  APNEA_INDEX: 'Índice de apnea',
+  CHEYNE_STOKES: 'Cheyne-Stokes',
+  OXYGEN_DESATURATION: 'Desaturación de oxígeno',
+  OXYGEN_SATURATION_PERCENTAGE: '% de saturación de oxígeno',
+  OXYGEN_SATURATION_EVAL_TIME_PERCENTAGE: '% satur. oxíg. - tiempo estudio',
+  BREATHS: 'Respiraciones',
+  PULSE_BPM: 'Pulso - rpm',
+  ANALYSIS_GUIDELINES: 'Criterios análisis',
+  INTERPRETATION: 'Interpretation',
+  PRINTED_ON: 'Printed on',
+};
+
+module.exports = labels;
