+++ src/handlers/languages/english.js
@@ -1,0 +1,30 @@
+const labels = {
+  PATIENT_ID: 'Patient ID',
+  DOB: 'DOB',
+  AGE: 'Age',
+  GENDER: 'Gender',
+  BMI: 'BMI',
+  RECORDING_DETAILS: 'Recording details',
+  DEVICE: 'Device',
+  RECORDING: 'Recording',
+  MONITORING_TIME_FLOW: 'Monitoring time (flow)',
+  FLOW_EVALUATION: 'Flow evaluation',
+  OXYGEN_SATURATION_EVALUATION: 'Oxygen saturation evaluation',
+  EVENTS_INDEX: 'Events index',
+  SUPINE: 'Supine',
+  NON_SUPINE: 'Non-supine',
+  UPRIGHT: 'Upright',
+  EVENTS_TOTAL: 'Events totals',
+  APNEA_INDEX: 'Apnea Index',
+  CHEYNE_STOKES: 'Cheyne-Stokes respiration',
+  OXYGEN_DESATURATION: 'Oxygen desaturation',
+  OXYGEN_SATURATION_PERCENTAGE: 'Oxygen saturation %',
+  OXYGEN_SATURATION_EVAL_TIME_PERCENTAGE: 'Oxygen saturation - eval time %',
+  BREATHS: 'Breaths',
+  PULSE_BPM: 'Pulse - bpm',
+  ANALYSIS_GUIDELINES: 'Analysis guidelines',
+  INTERPRETATION: 'Interpretation',
+  PRINTED_ON: 'Printed on',
+};
+
+module.exports = labels;
