+++ src/handlers/languages/portuguese.js
@@ -1,0 +1,30 @@
+const labels = {
+  PATIENT_ID: 'ID Paciente',
+  DOB: 'Data Nascimento',
+  AGE: 'Idade',
+  GENDER: 'Sexo',
+  BMI: 'IMC',
+  RECORDING_DETAILS: 'A gravar detalhes',
+  DEVICE: 'Dispositivo',
+  RECORDING: 'A gravar',
+  MONITORING_TIME_FLOW: 'Avaliação do fluxo',
+  FLOW_EVALUATION: 'Flow evaluation',
+  OXYGEN_SATURATION_EVALUATION: 'Aval. satur. Oxigénio',
+  EVENTS_INDEX: 'Índice de eventos',
+  SUPINE: 'Decúbito dorsal',
+  NON_SUPINE: 'Decúbito ventral',
+  UPRIGHT: 'Vertical',
+  EVENTS_TOTAL: 'Eventos totais',
+  APNEA_INDEX: 'Índice de apneia',
+  CHEYNE_STOKES: 'Cheyne-Stokes',
+  OXYGEN_DESATURATION: 'Dessaturação de oxigénio',
+  OXYGEN_SATURATION_PERCENTAGE: '% saturação de oxigénio',
+  OXYGEN_SATURATION_EVAL_TIME_PERCENTAGE: 'Satur. oxigénio — % tempo aval. Sat.',
+  BREATHS: 'Respirações',
+  PULSE_BPM: 'Pulso — bpm',
+  ANALYSIS_GUIDELINES: 'Orient. análise',
+  INTERPRETATION: 'Interpretação',
+  PRINTED_ON: 'Impresso em',
+};
+
+module.exports = labels;
